Move ActivationConditionsToPatchMapEntries to activation_condition.h
